requirements.txt
pyaudio
requests
sounddevice
numpy
scipy
python-dotenv

